fix: Atom stores in different file
